Struct MtgTools::api::Card

pub struct Card {

    pub arena_id: Option<u64>,
    pub id: String,
    pub lang: String,
    pub mtgo_id: Option<u64>,
    pub mtgo_foil_id: Option<u64>,
    pub multiverse_ids: Option<Vec<u64>>,
    pub tcgplayer_id: Option<u64>,
    pub tcgplayer_etched_id: Option<u64>,
    pub cardmarket_id: Option<u64>,
    pub object: String,
    pub oracle_id: Option<String>,
    pub prints_search_uri: String,
    pub rulings_uri: String,
    pub scryfall_uri: String,
    pub uri: String,
    pub name: String,
    pub type_line: Option<String>,
    pub rarity: String,
    pub collector_number: String,
    pub set_name: String,
    pub set_search_uri: String,
    pub set_type: String,
    pub set_uri: String,
    pub set: String,
    pub set_id: String,

}

https://scryfall.com/docs/api/cards

Fields
arena_id: Option<u64>

This card’s Arena ID, if any. A large percentage of cards are not available on Arena and do not have this ID.

id: String

A unique ID for this card in Scryfall’s database.

lang: String

A language code for this printing.

mtgo_id: Option<u64>

This card’s Magic Online ID (also known as the Catalog ID), if any. A large percentage of cards are not available on Magic Online and do not have this ID.

mtgo_foil_id: Option<u64>

This card’s foil Magic Online ID (also known as the Catalog ID), if any. A large percentage of cards are not available on Magic Online and do not have this ID.

multiverse_ids: Option<Vec<u64>>

This card’s multiverse IDs on Gatherer, if any, as an array of integers. Note that Scryfall includes many promo cards, tokens, and other esoteric objects that do not have these identifiers.

tcgplayer_id: Option<u64>

This card’s ID on TCGplayer’s API, also known as the productId.

tcgplayer_etched_id: Option<u64>

This card’s ID on TCGplayer’s API, for its etched version if that version is a separate product.

cardmarket_id: Option<u64>

This card’s ID on Cardmarket’s API, also known as the idProduct.

object: String

A content type for this object, always card.

oracle_id: Option<String>

A unique ID for this card’s oracle identity. This value is consistent across reprinted card editions, and unique among different cards with the same name (tokens, Unstable variants, etc).

prints_search_uri: String

A link to where you can begin paginating all re/prints for this card on Scryfall’s API.

rulings_uri: String

A link to this card’s rulings list on Scryfall’s API.

scryfall_uri: String

A link to this card’s permapage on Scryfall’s website.

uri: String

A link to this card object on Scryfall’s API.

name: String

The name of this card. If this card has multiple faces, this field will contain both names separated by ␣//␣.

type_line: Option<String>

The type line of this card.

rarity: String

This card’s rarity. One of common, uncommon, rare, special, mythic, or bonus.

collector_number: String

This card’s collector number. Note that collector numbers can contain non-numeric characters, such as letters or ★.

set_name: String

This card’s full set name.

set_search_uri: String

A link to where you can begin paginating this card’s set on the Scryfall API.

set_type: String

The type of set this printing is in.

set_uri: String

A link to this card’s set object on Scryfall’s API.

set: String

This card’s set code.

set_id: String

This card’s Set object UUID.